refactor(runtime): share llm content conversion helpers
